+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
 ## v3.0.8
-Updated on 2025-Nov-15
+Updated on 2025-Nov-16
 *   NEW: Additional "Notification Event Type" to set up notifications and run custom user scripts when a ping test fails.
+*   NEW: Added "Ping Test Failure" events to the results shown on the WebUI page.
 *   IMPROVED: Miscellaneous code improvements.
 
 ## v3.0.7